requirements.txt
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
audioread==2.1.9
autopep8==1.5.6
bcrypt==3.2.0
boto==2.49.0
boto3==1.21.3
botocore==1.24.3
captcha==0.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.14
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-import-export==2.7.1
django-storages==1.12.3
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
gunicorn==20.0.4
idna==3.3
jmespath==0.10.0
joblib==1.1.0
librosa==0.9.1
llvmlite==0.38.0
MarkupPy==1.14
mysqlclient==2.1.0
numba==0.55.1
numpy==1.21.5
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.1
pooch==1.6.0
psycopg2==2.9.3
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pypng==0.0.21
PyQRCode==1.2.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
resampy==0.2.2
s3transfer==0.5.1
scikit-learn==1.0.2
scipy==1.8.0
sib-api-v3-sdk==7.5.0
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
tablib==3.2.0
threadpoolctl==3.1.0
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
wincertstore==0.2
xendit-python==0.1.7
xlrd==2.0.1
xlwt==1.3.0
